git-svn-id: svn://123.57.173.36/meijia-parent@1759 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/vo/xcloud/CompanyAdminSearchVo.java
@@ -17,7 +17,7 @@ public class CompanyAdminSearchVo{
 	
 	private String passMd5;
 		
-	private int isCreate;
+	private Short isCreate;
 	
 	private Long startTime;
 	
@@ -55,11 +55,11 @@ public class CompanyAdminSearchVo{
 		this.passMd5 = passMd5;
 	}
 
-	public int getIsCreate() {
+	public Short getIsCreate() {
 		return isCreate;
 	}
 
-	public void setIsCreate(int isCreate) {
+	public void setIsCreate(Short isCreate) {
 		this.isCreate = isCreate;
 	}
